Update expander (#61)
* update expander version

* update expander to current dev

* add default gkr config (changes from #58)
