+++ Cargo.toml
@@ -20,6 +20,8 @@ halo2curves = { git = "https://github.com/PolyhedraZK/halo2curves", default-feat
     "bits",
 ] }
 arith = { git = "https://github.com/PolyhedraZK/Expander", branch = "dev" }
+mpi_config = { git = "https://github.com/PolyhedraZK/Expander", branch = "dev" }
+gkr_field_config = { git = "https://github.com/PolyhedraZK/Expander", branch = "dev" }
 expander_config = { git = "https://github.com/PolyhedraZK/Expander", branch = "dev", package = "config" }
 expander_circuit = { git = "https://github.com/PolyhedraZK/Expander", branch = "dev", package = "circuit" }
 gkr = { git = "https://github.com/PolyhedraZK/Expander", branch = "dev" }
