+++ expander_compiler/Cargo.toml
@@ -12,6 +12,8 @@ clap.workspace = true
 ethnum.workspace = true
 halo2curves.workspace = true
 tiny-keccak.workspace = true
+mpi_config.workspace = true
+gkr_field_config.workspace = true
 expander_config.workspace = true
 expander_circuit.workspace = true
 gkr.workspace = true
