+++ expander_compiler/src/circuit/layered/export.rs
@@ -2,7 +2,7 @@ use super::*;
 
 impl<C: Config> Circuit<C, NormalInputType> {
     pub fn export_to_expander<
-        DestConfig: expander_config::GKRConfig<CircuitField = C::CircuitField>,
+        DestConfig: gkr_field_config::GKRFieldConfig<CircuitField = C::CircuitField>,
     >(
         &self,
     ) -> expander_circuit::RecursiveCircuit<DestConfig> {
